reverting datatype change to be able to release this change before version 3.0

+++ src/main/java/redis/clients/jedis/BinaryClient.java
@@ -36,7 +36,7 @@ public class BinaryClient extends Connection {
 
   private String password;
 
-  private int db;
+  private long db;
 
   private boolean isInWatch;
 
@@ -73,7 +73,7 @@ public class BinaryClient extends Connection {
     this.password = password;
   }
   
-  public void setDb(int db) {
+  public void setDb(long db) {
     this.db = db;
   }
 
@@ -931,7 +931,7 @@ public class BinaryClient extends Connection {
     sendCommand(GETRANGE, key, toByteArray(startOffset), toByteArray(endOffset));
   }
 
-  public int getDB() {
+  public Long getDB() {
     return db;
   }
