+++ src/main/java/redis/clients/jedis/BinaryJedis.java
@@ -2877,7 +2877,7 @@ public class BinaryJedis implements BasicCommands, BinaryJedisCommands, MultiKey
     }
   }
 
-  public int getDB() {
+  public Long getDB() {
     return client.getDB();
   }
