+++ src/test/java/redis/clients/jedis/tests/JedisPoolTest.java
@@ -173,16 +173,16 @@ public class JedisPoolTest extends Assert {
         "foobared");
 
     Jedis jedis0 = pool.getResource();
-    assertEquals(0, jedis0.getDB());
+    assertEquals(0L, jedis0.getDB().longValue());
 
     jedis0.select(1);
-    assertEquals(1, jedis0.getDB());
+    assertEquals(1L, jedis0.getDB().longValue());
 
     pool.returnResource(jedis0);
 
     Jedis jedis1 = pool.getResource();
     assertTrue("Jedis instance was not reused", jedis1 == jedis0);
-    assertEquals(0, jedis1.getDB());
+    assertEquals(0L, jedis1.getDB().longValue());
 
     pool.returnResource(jedis1);
     pool.destroy();
